+++ tests/integration.rs
@@ -26,7 +26,7 @@ mod cli {
             pub global: GlobalOptions,
 
             #[command(subcommand)]
-            pub command: Commands,
+            pub command: Option<Commands>,
         }
 
         #[derive(Subcommand, Debug)]